+++ src/Models/Forms/AccountCredentialsForm.php
@@ -116,6 +116,7 @@ class AccountCredentialsForm extends Model
 
         if ($this->newPassword) {
             $this->user->afterPasswordChange();
+            Yii::$app->getUser()->destroyOtherSessions($this->user);
         }
     }